+++ delphin/mrs/query.py
@@ -1,4 +1,10 @@
+
+import warnings
 from itertools import product
+
+from delphin.mrs.components import nodes, links, var_id
+from delphin.mrs.util import rargname_sortkey
+from delphin.mrs.config import IVARG_ROLE
 
 # query methods
 def select_nodeids(xmrs, iv=None, label=None, pred=None):
@@ -7,16 +13,12 @@ def select_nodeids(xmrs, iv=None, label=None, pred=None):
     matching *iv* (intrinsic variable), *label*, or *pred* values. If
     none match, return an empty list.
     """
-    g = xmrs._graph
-    nids = []
-    datamatch = lambda d: ((iv is None or d['iv'] == iv) and
-                           (pred is None or d['pred'] == pred) and
-                           (label is None or d['label'] == label))
-    for nid in g.nodeids:
-        data = g.node[nid]
-        if datamatch(data):
-            nids.append(nid)
-    return nids
+    def datamatch(nid):
+        ep = xmrs.ep(nid)
+        return ((iv is None or ep.iv == iv) and
+                (pred is None or ep.pred == pred) and
+                (label is None or ep.label == label))
+    return list(filter(datamatch, xmrs.nodeids()))
 
 
 def select_nodes(xmrs, nodeid=None, pred=None):
@@ -26,7 +28,7 @@ def select_nodes(xmrs, nodeid=None, pred=None):
     """
     nodematch = lambda n: ((nodeid is None or n.nodeid == nodeid) and
                            (pred is None or n.pred == pred))
-    return list(filter(nodematch, xmrs.nodes))
+    return list(filter(nodematch, nodes(xmrs)))
 
 
 def select_eps(xmrs, nodeid=None, iv=None, label=None, pred=None):
@@ -39,34 +41,42 @@ def select_eps(xmrs, nodeid=None, iv=None, label=None, pred=None):
                          (iv is None or n.iv == iv) and
                          (label is None or n.label == label) and
                          (pred is None or n.pred == pred))
-    return list(filter(epmatch, xmrs.eps))
+    return list(filter(epmatch, xmrs.eps()))
 
 
 def select_args(xmrs, nodeid=None, rargname=None, value=None):
     """
-    Return the list of all |Arguments| that have the matching
-    *nodeid*, *rargname*, and/or *value* values. If none match,
+    Return a list of triples (nodeid, rargname, value) for all arguments
+    matching *nodeid*, *rargname*, and/or *value* values. If none match,
     return an empty list.
     """
-    argmatch = lambda a: ((nodeid is None or a.nodeid == nodeid) and
+    argmatch = lambda a: ((nodeid is None or a[0] == nodeid) and
                           (rargname is None or
-                           a.rargname.upper() == rargname.upper()) and
-                          (value is None or a.value == value))
-    return list(filter(argmatch, xmrs.args))
+                           a[1].upper() == rargname.upper()) and
+                          (value is None or a[2] == value))
+    all_args = (
+        (nid, role, val)
+        for nid in xmrs.nodeids()
+        for role, val in sorted(
+            xmrs.args(nid).items(),
+            key=lambda i: rargname_sortkey(i[0])
+        )
+    )
+    return list(filter(argmatch, all_args))
 
 
-def select_links(xmrs, source=None, target=None, rargname=None, post=None):
+def select_links(xmrs, start=None, end=None, rargname=None, post=None):
     """
-    Return the list of all |Links| that have the matching *source*,
-    *target*, *rargname*, and/or *post* values. If none match, return
+    Return the list of all |Links| that have the matching *start*,
+    *end*, *rargname*, and/or *post* values. If none match, return
     an empty list.
     """
     linkmatch = lambda l: (
-        (source is None or l.source == source) and
-        (target is None or l.target == target) and
+        (start is None or l.start == start) and
+        (end is None or l.end == end) and
         (rargname is None or l.rargname == rargname) and
         (post is None or l.post == post))
-    return list(filter(linkmatch, xmrs.links))
+    return list(filter(linkmatch, links(xmrs)))
 
 
 def select_hcons(xmrs, hi=None, relation=None, lo=None):
@@ -79,7 +89,7 @@ def select_hcons(xmrs, hi=None, relation=None, lo=None):
         (hi is None or hc.hi == hi) and
         (relation is None or hc.relation == relation) and
         (lo is None or hc.lo == lo))
-    return list(filter(hcmatch, xmrs.hcons))
+    return list(filter(hcmatch, xmrs.hcons()))
 
 
 def select_icons(xmrs, left=None, relation=None, right=None):
@@ -92,7 +102,7 @@ def select_icons(xmrs, left=None, relation=None, right=None):
         (left is None or ic.left == left) and
         (relation is None or ic.relation == relation) and
         (right is None or ic.right == right))
-    return list(filter(icmatch, xmrs.icons))
+    return list(filter(icmatch, xmrs.icons()))
 
 
 def find_argument_target(xmrs, nodeid, rargname):
@@ -123,80 +133,24 @@ def find_argument_target(xmrs, nodeid, rargname):
         will be returned. With this nodeid, one can then use
         :py:meth:`find_quantifier` to get its quantifier's nodeid.
     """
-    g = xmrs._graph
-    tgt = None
-    try:
-        tgt_val = xmrs.get_arg(nodeid, rargname).value
-        tgt_attr = g.node[tgt_val]
+    tgt = xmrs.args(nodeid)[rargname]
+    if tgt in xmrs.variables():
+        try:
+            return xmrs.nodeid(tgt)
+        except KeyError:
+            pass
 
-        # intrinsic variable
-        if 'iv' in tgt_attr:
-            tgt = tgt_attr['iv']
+        try:
+            tgt = xmrs.hcon(tgt).lo
+            return next(iter(xmrs.labelset_heads(tgt)), None)
+        except KeyError:
+            pass
 
-        # hcons; tgt_val is a hole
-        if 'hcons' in tgt_attr:
-            tgt_val = tgt_attr['hcons'].lo
-        # label or hcons lo variable (see previous if block)
-        if tgt_val in g.labels:
-            tgt = xmrs.labelset_head(tgt_val)
-
-        # otherwise likely a constant or unbound variable
-        tgt = tgt_val
-
-    # nodeid or rargname were missing, or tgt_val wasn't a node
-    except (AttributeError, KeyError):
-        pass
-        #logging.warning('Cannot find argument target; argument is '
-        #                'invalid: {}:{}'.format(nodeid, rargname))
+        try:
+            return next(iter(xmrs.labelset_heads(tgt)))
+        except (KeyError, StopIteration):
+            pass
     return tgt
-
-
-def find_quantifier(xmrs, nodeid):
-    """
-    Return the nodeid of the quantifier of the EP given by `nodeid`.
-
-    Args:
-        xmrs: The |Xmrs| object to use.
-        nodeid: The nodeid of the quantified EP/node.
-    Returns:
-        The nodeid of the quantifier for `nodeid`. If `nodeid` is not
-        in the Xmrs, it itself is a quantifier, or if it does not have
-        a quantifier, None is returned.
-    """
-    ep = xmrs.get_ep(nodeid)
-    if (not ep or
-        ep.is_quantifier() or
-        ep.iv not in xmrs._graph.node or
-        'bv' not in xmrs._graph.node[ep.iv]):
-        # in some subgraphs, an IV might not exist even when specified
-        return None
-    return xmrs._graph.node[ep.iv]['bv']
-
-
-def get_outbound_args(xmrs, nodeid, allow_unbound=True):
-    """
-    Yield the |Arguments| of `nodeid` that point to other EPs/nodes.
-
-    Args:
-        xmrs: The |Xmrs| object to use.
-        nodeid: The nodeid of the EP/node whose arguments to yield.
-        allow_unbound: If True, also yield arguments that point to
-            unbound (e.g. dropped) EPs/nodes or constants.
-    Yields:
-        |Arguments| whose targets are not the given `nodeid`.
-    """
-    g = xmrs._graph
-    ep = xmrs.get_ep(nodeid)
-    for arg in ep.args.values():
-        nid = arg.nodeid
-        tgt = arg.value
-        data = g.node.get(tgt, {})
-        # ignore intrinsic arguments
-        if data.get('iv') == nid or data.get('bv') == nid:
-            continue
-        is_outbound = 'iv' in data or 'hcons' in data or tgt in g.labels
-        if (allow_unbound or is_outbound):
-            yield arg
 
 
 def find_subgraphs_by_preds(xmrs, preds, connected=None):
@@ -214,17 +168,20 @@ def find_subgraphs_by_preds(xmrs, preds, connected=None):
         |Xmrs| objects for the found subgraphs.
     """
     preds = list(preds)
+    count = len(preds)
     # find all lists of nodeids such that the lists have no repeated nids;
     # keep them as a list (fixme: why not just get sets?)
-    nidsets = list(
-        filter(lambda ps: len(set(ps)) == len(ps),
-               map(lambda p: select_nodeids(xmrs, pred=p), preds))
+    nidsets = set(
+        tuple(sorted(ns))
+        for ns in filter(
+            lambda ns: len(set(ns)) == count,
+            product(*[select_nodeids(xmrs, pred=p) for p in preds])
+        )
     )
-    for sg in map(xmrs.subgraph, product(*nidsets)):
-        if connected is not None and sg.is_connected() != connected:
-            continue
-        yield sg
-
+    for nidset in nidsets:
+        sg = xmrs.subgraph(nidset)
+        if connected is None or sg.is_connected() == connected:
+            yield sg
 
 
 # def introduced_variables(xmrs):
@@ -235,32 +192,18 @@ def intrinsic_variable(xmrs, nid):
     return xmrs.args(nid).get(IVARG_ROLE, None)
 
 def intrinsic_variables(xmrs):
-    return [ep[3][IVARG_ROLE] for ep in xmrs.eps()
-            if not ep[1].is_quantifier() and IVARG_ROLE in ep[3]]
+    ivs = set(
+        ep[3][IVARG_ROLE] for ep in xmrs.eps()
+        if not ep[1].is_quantifier() and IVARG_ROLE in ep[3]
+    )
+    return sorted(ivs, key=var_id)
 
 def bound_variables(xmrs):
-    return [ep[3][IVARG_ROLE] for ep in xmrs.eps()
-            if ep[1].is_quantifier() and IVARG_ROLE in ep[3]]
-
-
-def nodeid(xmrs, iv, quantifier=False):
-    """
-    Retrieve the nodeid of an |EP| given an intrinsic variable.
-
-    Args:
-        iv: The intrinsic variable of the |EP|.
-        quantifier: If True and `iv` is the bound variable of a
-            quantifier, return the nodeid of the quantifier. False
-            by default.
-    Returns:
-        An integer nodeid.
-    """
-    nids = xmrs._vars[iv]['refs'].get(IVARG_ROLE, [])
-    # return the nid only
-    for nid in nids:
-        if xmrs.ep(nid)[1].is_quantifier() == quantifier:
-            return nid
-    return None  # raise error?
+    bvs = set(
+        ep[3][IVARG_ROLE] for ep in xmrs.eps()
+        if ep[1].is_quantifier() and IVARG_ROLE in ep[3]
+    )
+    return sorted(bvs, key=var_id)
 
 def in_labelset(xmrs, nodeids, label=None):
     """
@@ -274,8 +217,47 @@ def in_labelset(xmrs, nodeids, label=None):
     """
     nodeids = set(nodeids)
     if label is None:
-        label = xmrs._eps[next(iter(nodeids))][2]
+        label = xmrs.ep(next(iter(nodeids))).label
     return nodeids.issubset(xmrs._vars[label]['refs']['LBL'])
 
 
 #def audit(xmrs): inspect well-formedness and report individual errors
+
+# deprecated
+
+def find_quantifier(xmrs, nodeid):
+    warnings.warn(
+        'find_quantifier() is deprecated; '
+        'try xmrs.nodeid(xmrs.ep(nodeid).iv, quantifier=True)',
+        DeprecationWarning
+    )
+    ep = xmrs.ep(nodeid)
+    if not ep.is_quantifier():
+        return xmrs.nodeid(xmrs.ep(nodeid).iv, quantifier=True)
+    return None
+
+def get_outbound_args(xmrs, nodeid, allow_unbound=True):
+    warnings.warn(
+        'get_outbound_args() is deprecated; '
+        'try xmrs.outgoing_args(nodeid)',
+        DeprecationWarning
+    )
+    tgts = set(intrinsic_variables(xmrs))
+    tgts.update(xmrs.labels())
+    tgts.update(hc.hi for hc in xmrs.hcons())
+    return [
+        (nodeid, role, val)
+        for role, val in sorted(
+            xmrs.outgoing_args(nodeid).items(),
+            key=lambda r_v: rargname_sortkey(r_v[0])
+        )
+        if allow_unbound or val in tgts
+    ]
+
+def nodeid(xmrs, iv, quantifier=False):
+    warnings.warn(
+        'nodeid() is deprecated; '
+        'try xmrs.nodeid(iv, quantifier=quantifier)',
+        DeprecationWarning
+    )
+    return xmrs.nodeid(iv, quantifier=quantifier)